had to add features to x86

+++ 91/src/gdt.rs
@@ -10,11 +10,11 @@ static mut TSS: x86_64::structures::tss::TaskStateSegment =
     x86_64::structures::tss::TaskStateSegment::new();
 static mut STACK: [u8; STACK_SIZE] = [0; STACK_SIZE];
 
-pub const DOUBLE_FAULT_IST_INDEX: usize = 0;
+pub const DOUBLE_FAULT_IST_INDEX: u16 = 0;
 
 pub fn init_gdt() {
     unsafe {
-        TSS.interrupt_stack_table[DOUBLE_FAULT_IST_INDEX] =
+        TSS.interrupt_stack_table[DOUBLE_FAULT_IST_INDEX as usize] =
             x86_64::VirtAddr::from_ptr(&raw const STACK) + STACK_SIZE;
         let kcs = GDT.add_entry(x86_64::structures::gdt::Descriptor::kernel_code_segment());
         let _kds = GDT.add_entry(x86_64::structures::gdt::Descriptor::user_data_segment());
